Deleted common and misc
