--- fenicsxconcrete/experimental_setup/common.py
@@ -1,13 +1,0 @@
-"""Based on Philipp Diercks implementation for multi"""
-"""common helpers"""
-
-
-def to_floats(values):
-    floats = []
-    try:
-        for v in values:
-            floats.append(float(v))
-    except TypeError as e:
-        floats = [float(values)]
-
-    return floats
